test: split reset concurrency coverage
