+++ src/gateway/server.sessions.reset-cleanup.test.ts
@@ -11,7 +11,7 @@ import {
   registerAgentHarness,
   restoreRegisteredAgentHarnesses,
 } from "../agents/harness/registry.js";
-import { loadSessionEntry, loadTranscriptEvents } from "../config/sessions/session-accessor.js";
+import { loadSessionEntry } from "../config/sessions/session-accessor.js";
 import type { SessionAcpMeta } from "../config/sessions/types.js";
 import { enqueueSystemEvent, peekSystemEvents } from "../infra/system-events.js";
 import {
@@ -38,7 +38,6 @@ import {
   getGatewayConfigModule,
   getSessionsHandlers,
   sessionHookMocks,
-  sessionLifecycleHookMocks,
 } from "./test/server-sessions.test-helpers.js";
 
 const { createSessionStoreDir, seedActiveMainSession } = setupGatewaySessionsTestHarness();
@@ -579,75 +578,6 @@ test("sessions.reset finishes after lifecycle rotation during destructive cleanu
   expect(reset.ok).toBe(true);
   expectResetAcpState(readAcpSessionMeta({ sessionKey: "agent:main:main" }));
   expect(prepareFreshSession).not.toHaveBeenCalled();
-});
-
-test("sessions.reset preserves a concurrent same-id lifecycle replacement", async () => {
-  const registeredHarnesses = listRegisteredAgentHarnesses();
-  const { storePath } = await createSessionStoreDir();
-  await writeSessionStore({
-    entries: {
-      main: sessionStoreEntry("sess-main", {
-        lifecycleRevision: "original-revision",
-      }),
-    },
-  });
-  let lifecycleCurrent = true;
-  registerAgentHarness({
-    id: "reset-race-observer",
-    label: "Reset race observer",
-    supports: () => ({ supported: false }),
-    runAttempt: async () => {
-      throw new Error("not used");
-    },
-    reset: async () => {
-      lifecycleCurrent = false;
-      await writeSessionStore({
-        entries: {
-          main: sessionStoreEntry("sess-main", {
-            label: "newer owner",
-            lifecycleRevision: "replacement-revision",
-          }),
-        },
-      });
-    },
-  });
-  const { performGatewaySessionReset } = await import("./session-reset-service.js");
-
-  try {
-    const reset = await performGatewaySessionReset({
-      key: "main",
-      reason: "new",
-      commandSource: "gateway:agent",
-      assertCurrent: () => {
-        if (!lifecycleCurrent) {
-          throw new Error("stale lifecycle");
-        }
-      },
-    });
-
-    expect(reset).toMatchObject({
-      ok: true,
-      entry: {
-        label: "newer owner",
-        lifecycleRevision: "replacement-revision",
-        sessionId: "sess-main",
-      },
-    });
-    expect(
-      await loadTranscriptEvents({
-        agentId: "main",
-        sessionId: "sess-main",
-        sessionKey: "agent:main:main",
-        storePath,
-      }),
-    ).toEqual([]);
-    expect(sessionLifecycleHookMocks.runSessionEnd).not.toHaveBeenCalled();
-    expect(sessionLifecycleHookMocks.runSessionStart).not.toHaveBeenCalled();
-    expect(subagentLifecycleHookMocks.runSubagentEnded).not.toHaveBeenCalled();
-    expect(threadBindingMocks.unbindThreadBindingsBySessionKey).not.toHaveBeenCalled();
-  } finally {
-    restoreRegisteredAgentHarnesses(registeredHarnesses);
-  }
 });
 
 test("sessions.reset rejects a concurrent archive during lifecycle rotation", async () => {
